Update WaitingRoomControl/Panel, GameServer/Client, ServerGUI
Create a DataNeededForClientObject to be sent to the client whenever data pertaining to the game needs to be updated client side. It's not the most organized at the moment but it's to prevent us from sending the entire GameManager over... We shoud probably have a like GameData that is connected with GameManager to separate the logic from the jumble of data. Was going to use that to sent to Client but the Client shouldn't have access to all of the data that is in there.